+++ ConcreteEngine.Graphics/IMesh.cs
@@ -8,8 +8,8 @@ namespace ConcreteEngine.Graphics;
 
 public interface IMesh : IGraphicsResource
 {
-    public int VertexBufferId { get; }
-    public int IndexBufferId { get; }
+    public ushort VertexBufferId { get; }
+    public ushort IndexBufferId { get; }
     public ICollection<VertexAttributeDescriptor> VertexAttributes { get; }
     public bool HasIndices { get; }
     public uint DrawCount { get; set; }
